Fix build warnings
It has been a while since last successful build. There are various
warnings and errors from the building. Lets fix it first so that
we can improve the logics later.

Besides, the bitmap and rb_erase() are improved greatly. With all
these changes applied, the KVM virtual machine and vCPUs can be
created successfully. It gives me the start point to work on the
ELF loading.

Signed-off-by: Gavin Shan <gshan@redhat.com>